polymarket-clob-bot/requirements.txt
py-clob-client>=0.9.0
websockets>=12.0
aiohttp>=3.9
python-telegram-bot>=20.0
pandas>=2.0
numpy>=1.24
xgboost>=2.0
lightgbm>=4.0
scikit-learn>=1.3
matplotlib>=3.7
web3>=6.0
python-dotenv>=1.0
